fix: deletewarning working with i18n for now with extra cardWithImageMyShop.tsx
