+++ components/myShop/myshopContent.tsx
@@ -1,4 +1,5 @@
 import { CardWithImageMyShop } from '../ui/cardWithImageMyShop'
+import { CardWithImage } from '../ui/cardWithImage'
 import { getProductsOwned } from '@/lib/product-actions'
 import { getUser } from '@/lib/user-actions'
 import AnimatedCard from '../ui/animated-card'
@@ -20,12 +21,12 @@ const MyShopContent = async () => {
         <div className="-mx-2 mt-[20px] flex flex-wrap justify-around overflow-y-hidden">
           {products?.map((p, index) => (
             // <AnimatedCard  delay={0.3} >
-              <CardWithImageMyShop
-                key={`pr-${index}`}
-                className="mx-[5px] my-[0.5rem]"
-                product={products[index]}
-                editable
-              />
+            <CardWithImageMyShop
+              key={`pr-${index}`}
+              className="mx-[5px] my-[0.5rem]"
+              product={products[index]}
+              editable
+            />
             // </AnimatedCard>
           ))}
         </div>
