+++ package.json
@@ -56,7 +56,7 @@
     "framer-motion": "^11.2.11",
     "lucide-react": "^0.383.0",
     "nanoid": "^5.0.7",
-    "next": "^14.2.3",
+    "next": "^14.2.4",
     "next-auth": "^5.0.0-beta.17",
     "next-intl": "^3.12.2",
     "next-themes": "^0.3.0",
